# packages/sie_gateway/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignedWorker": {
        "properties": {
          "bundle": {
            "type": "string"
          },
          "gpu": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "gpu"
        ],
        "type": "object"
      },
...
      "ConfigModelDocument": {
        "properties": {
          "bundles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "sie_id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "sie_id",
          "source",
          "bundles"
        ],
        "type": "object"
      },
      "ConfigModelSummary": {
        "properties": {
          "model_id": {
            "type": "string"
          },
          "profiles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "profiles",
          "source"
        ],
        "type": "object"
      },
      "ConfigModelsResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ConfigModelSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "models"
        ],
        "type": "object"
      },
      "CreatePoolRequest": {
        "anyOf": [
          {
            "properties": {
              "gpus": {
                "minProperties": 1
              }
            },
            "required": [
              "gpus"
            ]
          },
          {
            "properties": {
              "gpu_caps": {
                "minProperties": 1
              }
            },
            "required": [
              "gpu_caps"
            ]
          }
        ],
        "properties": {
          "bundle": {
            "type": [
              "string",
              "null"
            ]
          },
          "gpu_caps": {
            "additionalProperties": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "gpus": {
            "additionalProperties": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "minimum_worker_count": {
            "description": "Per-pool warm floor (minimum machines kept warm via KEDA). Default 0\nkeeps scale-from-zero. See `PoolSpec::minimum_worker_count`.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "name": {
            "description": "Pool name used in gpu=\"pool/machine_profile\" routing. Names are stored and routed in lowercase. Only ASCII letters, digits, '_' and '-' are allowed; '_default' is reserved.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^(?!_[dD][eE][fF][aA][uU][lL][tT]$)[A-Za-z0-9_-]+$",
            "type": "string"
          },
          "pinned_models": {
            "description": "Per-pool pinned-model set. Each id is validated against the models the\ngateway already tracks and stored canonicalized; ids may be\nprofile-qualified (`model-name:profile_name`). Default empty leaves\nlazy-loading unchanged. See `PoolSpec::pinned_models`.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "queue_pool": {
            "description": "Physical Helm/NATS queue namespace to draw workers from. Omit for normal\ndynamic logical pools; they use `default`. Set only to an\noperator-provisioned Helm queue pool declared under\n`queueRouting.staticQueuePools` for dedicated capacity.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^(?!_[dD][eE][fF][aA][uU][lL][tT]$)[A-Za-z0-9_-]+$",
            "type": [
              "string",
              "null"
            ]
          },
          "ttl_seconds": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "ErrorDetailCore": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "Pool": {
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/PoolSpec"
          },
          "status": {
            "$ref": "#/components/schemas/PoolStatus"
          }
        },
        "required": [
          "spec",
          "status"
        ],
        "type": "object"
      },
      "PoolListResponse": {
        "properties": {
          "pools": {
            "items": {
              "$ref": "#/components/schemas/Pool"
            },
            "type": "array"
          }
        },
        "required": [
          "pools"
        ],
        "type": "object"
      },
...
      "PoolStatus": {
        "properties": {
          "assigned_workers": {
            "items": {
              "$ref": "#/components/schemas/AssignedWorker"
            },
            "type": "array"
          },
          "created_at": {
            "format": "double",
            "type": "number"
          },
          "last_renewed": {
            "format": "double",
            "type": "number"
          },
          "state": {
            "$ref": "#/components/schemas/PoolState"
          }
        },
        "required": [
          "state"
        ],
        "type": "object"
      },
...
      "StandardApiError": {
        "description": "FastAPI-style error envelope for most gateway JSON errors.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetailCore"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "SIE bearer token. Mutating pool/config/admin routes require the configured admin token; other protected routes accept a normal gateway token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Rust gateway API for SIE inference, pool coordination, and read-only runtime config.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "SIE Gateway",
    "version": "0.6.27"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/configs/models": {
      "get": {
        "operationId": "get_model_configs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigModelsResponse"
                }
              }
            },
            "description": "Model configs visible to this gateway replica"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Gateway auth enabled but no tokens configured"
          }
        },
        "summary": "GET /v1/configs/models - List all model configs visible to this gateway.",
        "tags": [
          "config"
        ]
      }
    },
    "/v1/configs/models/{id}": {
      "get": {
        "description": "Handles two endpoints sharing the same wildcard route:\n\n- `GET /v1/configs/models/{id}` — YAML view of the gateway's model.\n- `GET /v1/configs/models/{id}/status` — JSON worker-ack readiness for\n  that model on this specific gateway replica.\n\nThe `{*id}` wildcard is necessary because model IDs routinely contain\nforward slashes (`BAAI/bge-m3`), so we strip an optional trailing\n`/status` segment here rather than maintain two overlapping axum routes\n(axum's `matchit` panics on overlapping catch-all routes).\n\nDisambiguation when a model literally named `something/status` exists:\nwe prefer the status interpretation only when the `/status`-stripped\nid resolves to a real model. That keeps an accidental model whose ID\nends with `/status` addressable under `GET\n/v1/configs/models/foo/status` as a config read, and matches what\nadmin tooling expects (status requests are only meaningful against\nmodels that actually exist). `sie-config` additionally refuses to\nregister model IDs ending in `/status` so the pathological case where\n**both** `foo` and `foo/status` are models cannot arise, but we\nstill degrade gracefully if it ever does.",
        "operationId": "get_model_config_or_status",
        "parameters": [
          {
            "description": "Model id. Runtime route is Axum catch-all `/v1/configs/models/{*id}`; clients using this OpenAPI path template should percent-encode slashes in model ids, for example `BAAI%2Fbge-m3`.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/x-yaml": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigModelDocument"
                }
              }
            },
            "description": "Model config YAML"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Model not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Failed to serialize YAML; gateway auth enabled but no tokens configured"
          }
        },
        "summary": "GET /v1/configs/models/{*id} — dual-purpose dispatcher.",
        "tags": [
          "config"
        ],
        "x-sie-axum-catch-all": "/v1/configs/models/{*id}"
      }
    },
...
    "/v1/pools": {
      "get": {
        "operationId": "list_pools",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolListResponse"
                }
              }
            },
            "description": "Pools visible to this gateway replica"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Gateway auth enabled but no tokens configured"
          }
        },
        "tags": [
          "pools"
        ]
      },
      "post": {
        "operationId": "create_pool",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            },
            "description": "Pool created, renewed, or updated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Invalid pool request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Valid bearer token but admin token required for this mutation (or admin token not configured)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Gateway auth enabled but no tokens configured"
          }
        },
        "tags": [
          "pools"
        ]
      }
    },
    "/v1/pools/{name}": {
      "delete": {
        "operationId": "delete_pool",
        "parameters": [
          {
            "description": "Pool name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Pool deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Pool cannot be deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Pool not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Gateway auth enabled but no tokens configured"
          }
        },
        "tags": [
          "pools"
        ]
      },
      "get": {
        "operationId": "get_pool",
        "parameters": [
          {
            "description": "Pool name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            },
            "description": "Pool detail"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Pool not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Gateway auth enabled but no tokens configured"
          }
        },
        "tags": [
          "pools"
        ]
      }
    },
    "/v1/pools/{name}/renew": {
      "post": {
        "operationId": "renew_pool",
        "parameters": [
          {
            "description": "Pool name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Pool renewed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token (inference token)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Valid bearer token but admin token required for this mutation (or admin token not configured)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Pool not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Gateway auth enabled but no tokens configured"
          }
        },
        "tags": [
          "pools"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}